fix: content size bigint migration

+++ alembic/versions/2026_03_04_2252-8ca57ddc101c_change_content_size_to_bigint.py
@@ -23,8 +23,8 @@ def upgrade() -> None:
     op.alter_column(
         'item_metainfo',
         'content_size',
-        existing_type=sa.BIGINT(),
-        type_=sa.Integer(),
+        existing_type=sa.Integer(),
+        type_=sa.BIGINT(),
         existing_nullable=True,
     )
     # ### end Alembic commands ###
@@ -36,8 +36,8 @@ def downgrade() -> None:
     op.alter_column(
         'item_metainfo',
         'content_size',
-        existing_type=sa.Integer(),
-        type_=sa.BIGINT(),
+        existing_type=sa.BIGINT(),
+        type_=sa.Integer(),
         existing_nullable=True,
     )
     # ### end Alembic commands ###
